refactor: reorganize directory structure with protocols/ and app/ dirs
Move OpenDroneID and WiFi protocol files to src/protocols/ and main.cpp
to src/app/ for clearer separation of concerns. Update all include paths,
build configs, CI, and test wrappers accordingly.
