+++ .github/workflows/ci.yml
@@ -61,8 +61,8 @@ jobs:
         run: |
           python -m gcovr --root . --txt --print-summary \
             --fail-under-line 35 \
-            --filter src/opendroneid.c \
-            --filter src/wifi.c \
+            --filter src/protocols/opendroneid.c \
+            --filter src/protocols/wifi.c \
             --filter src/modules/detector_logic.cpp \
             --filter src/modules/flockyou_logic.cpp \
             --filter src/modules/foxhunter_logic.cpp \
@@ -137,4 +137,4 @@ jobs:
           cppcheck --quiet --error-exitcode=1 \
             --enable=warning,performance \
             --inline-suppr --suppress=missingIncludeSystem \
-            -I src src/main.cpp src/hal src/modules src/storage src/web src/opendroneid.c src/wifi.c
+            -I src -I src/protocols src/app src/hal src/modules src/storage src/web src/protocols
