+++ platformio.ini
@@ -16,6 +16,7 @@ build_flags =
     -DBOARD_HAS_PSRAM
     -mfix-esp32-psram-cache-issue
     -DCONFIG_BT_NIMBLE_ENABLED=1
+    -Isrc/protocols
 
 ; Upload options
 upload_speed = 921600
@@ -45,13 +46,12 @@ build_src_filter = +<*>
 ; Static analysis
 check_tool = cppcheck
 check_src_filters =
-    +<src/main.cpp>
+    +<src/app/>
     +<src/hal/>
     +<src/modules/>
     +<src/storage/>
     +<src/web/>
-    +<src/opendroneid.c>
-    +<src/wifi.c>
+    +<src/protocols/>
 check_flags =
     cppcheck: --enable=warning,performance,style,portability
     cppcheck: --inline-suppr
@@ -70,6 +70,7 @@ build_src_filter =
     +<modules/skyspy_logic.cpp>
 test_build_src = no
 build_flags =
+    -Isrc/protocols
     -fprofile-arcs
     -ftest-coverage
     -lgcov
@@ -82,6 +83,7 @@ build_src_filter =
     -<*>
 test_build_src = no
 build_flags =
+    -Isrc/protocols
     -fprofile-arcs
     -ftest-coverage
     -lgcov
