+++ test/test_native_parser_edge/opendroneid_under_test.c
@@ -1,1 +1,1 @@
-#include "../../src/opendroneid.c"
+#include "../../src/protocols/opendroneid.c"
