+++ test/test_native_parser_edge/wifi_under_test.c
@@ -1,1 +1,1 @@
-#include "../../src/wifi.c"
+#include "../../src/protocols/wifi.c"
