+++ test/test_native_smoke/opendroneid_under_test.c
@@ -1,1 +1,1 @@
-#include "../../src/opendroneid.c"
+#include "../../src/protocols/opendroneid.c"
