+++ test/test_native_smoke/wifi_under_test.c
@@ -1,1 +1,1 @@
-#include "../../src/wifi.c"
+#include "../../src/protocols/wifi.c"
